+++ releasenotes.md
@@ -20,6 +20,7 @@
 - Constdef declarations introduced.
 - Properly support `@deprecated` as contract.
 - Support deprecating enum values.
+- Improve error when trying to use an extern const as a compile time constant. #2969
 
 ### Stdlib changes
 - Summarize sort macros as generic function wrappers to reduce the amount of generated code. #2831